requirements.txt
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
xgboost>=2.0.0
pyod>=1.1.0
torch>=2.0.0
sentence-transformers>=2.2.0
streamlit>=1.20.0
joblib>=1.2.0
pyyaml>=6.0.0
pytest>=7.0.0